progress tooowards long double

+++ tests/57_macro_bug.c
@@ -4,13 +4,13 @@
 
 void main()
 {
-    float x = MIN(1,2);
+    double x = MIN(1,2);
     int y = 14;
-    float z;
+    double z;
     z = MIN(y, 13.5);
     y = MIN(y, 13);
     
-    float pi = 3.14;
+    double pi = 3.14;
     int pi_int = pi;
     
     printf("Macro test: %d %d %f %d \n", x, y, z, pi_int);
